migrate collecting training data example to gen2
