--- collecting-training-data/requirements.txt
@@ -1,3 +1,0 @@
-opencv-python==4.5.1.48
---extra-index-url https://artifacts.luxonis.com/artifactory/luxonis-python-snapshot-local/
-depthai==0.3.0.0+aeda4a9fdef6edc9f826b7dc354a123d1611a7c6
