+++ gen2-collecting-training-data/requirements.txt
@@ -1,0 +1,2 @@
+opencv-python
+depthai==2.1.0.0
